backend/requirements.txt
# Backend Dependencies for BulamuChainBot
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Authentication
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL (optional)

# AI and NLP Services
openai==1.3.3
google-generativeai==0.3.1
requests==2.31.0

# Speech Processing
SpeechRecognition==3.10.0
pyttsx3==2.90
google-cloud-speech==2.21.0
azure-cognitiveservices-speech==1.31.0

# Blockchain Integration
web3==6.11.3
eth-account==0.9.0

# Utilities
python-decouple==3.8
cryptography==41.0.7
Pillow==10.1.0
qrcode==7.4.2
python-multipart==0.0.6

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
black==23.10.1
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
